requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-environ==0.11.2
django-extensions==3.2.3
django-heroku==0.3.1
idna==3.7
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.1
whitenoise==6.7.0
